Correcting pypeline and fixing it
There is no guarantee that the code works but the pipeline passes so be careful and always push when the pupeline is green. If it's not green it's because your code is not correct not because mypy is dumb or pylint either. if you struggle executing the main try running it outside the package kdna.